#3700: Treat resource instances as resources.
The predicate that tells if a component class is "correctlyConfigured" only checked the 'resourceClasses' collection to determine whether the resource is actually a resource.
As a consequence a warning, was logged when an instantiated resource class was tested:

"WARNING: A provider war.HelloWorldResource registered in SERVER runtime does not implement any provider interfaces applicable in the SERVER runtime. Due to constraint configuration problems the provider war.HelloWorldResource will be ignored."

The fix is as conservative as possible.
Changed the result of the relevant predicate only in the context where the changed result values should not have any impact.
But the side-effect of logging a warning should go away.

No additional test since the change shouldn't have any observable effect other than the absence of a warning.