OutboundJaxrsResponse.from(): Don't propagate hop-by-hop headers
Without this PR, when you have jersey-server forwarding a request with
jersey-client to a remote service and there is an error response from
the remote service with `Transfer-Encoding: chunked` in the response
headers, jersey-server propagates the Transfer-Encoding header back to
its response. But jersey-server does not actually encode the response
using chunked encoding. Also it adds a Content-Length header, which is
also forbidden when a Transfer-Encoding header is present. So you end
up with a response similar to this:

    HTTP/1.1 500 Internal Server Error
    Server: Foo
    Transfer-Encoding: chunked
    Content-Length: 25

    { "error": "Some error" }

A client seeing such a response can easily break down, for example
with jersey-client itself (using Apache HttpComponents for transport):

```java
javax.ws.rs.client.ResponseProcessingException: Failed to convert a response into an exception.
	at org.glassfish.jersey.client.JerseyInvocation.convertToException(JerseyInvocation.java:1111)
	at org.glassfish.jersey.client.JerseyInvocation.translate(JerseyInvocation.java:948)
	at org.glassfish.jersey.client.JerseyInvocation.lambda$invoke$2(JerseyInvocation.java:787)
	at org.glassfish.jersey.internal.Errors.process(Errors.java:316)
	at org.glassfish.jersey.internal.Errors.process(Errors.java:298)
	at org.glassfish.jersey.internal.Errors.process(Errors.java:229)
	at org.glassfish.jersey.process.internal.RequestScope.runInScope(RequestScope.java:414)
	at org.glassfish.jersey.client.JerseyInvocation.invoke(JerseyInvocation.java:785)
	at org.glassfish.jersey.client.JerseyInvocation$Builder.method(JerseyInvocation.java:438)
	at org.glassfish.jersey.client.JerseyInvocation$Builder.get(JerseyInvocation.java:329)
	at Foo.main(Foo.java:43)
Caused by: javax.ws.rs.ProcessingException: Failed to buffer the message content input stream.
	at org.glassfish.jersey.message.internal.InboundMessageContext.bufferEntity(InboundMessageContext.java:931)
	at org.glassfish.jersey.client.InboundJaxrsResponse.bufferEntity(InboundJaxrsResponse.java:161)
	at org.glassfish.jersey.client.JerseyInvocation.convertToException(JerseyInvocation.java:1065)
	... 41 more
Caused by: org.apache.http.MalformedChunkCodingException: Bad chunk header: {
	at org.apache.http.impl.io.ChunkedInputStream.getChunkSize(ChunkedInputStream.java:277)
	at org.apache.http.impl.io.ChunkedInputStream.nextChunk(ChunkedInputStream.java:225)
	at org.apache.http.impl.io.ChunkedInputStream.read(ChunkedInputStream.java:184)
	at org.apache.http.conn.EofSensorInputStream.read(EofSensorInputStream.java:135)
	at java.io.BufferedInputStream.read1(BufferedInputStream.java:284)
	at java.io.BufferedInputStream.read(BufferedInputStream.java:345)
	at java.io.FilterInputStream.read(FilterInputStream.java:133)
	at java.io.FilterInputStream.read(FilterInputStream.java:107)
	at org.glassfish.jersey.message.internal.ReaderWriter.writeTo(ReaderWriter.java:116)
	at org.glassfish.jersey.message.internal.InboundMessageContext.bufferEntity(InboundMessageContext.java:918)
	... 43 more
```

The HTTP/1.1 spec says that connection-specific aka hop-by-hop headers
(including Transfer-Encoding) must not be forwarded, and the code
doing the forwarding in this case is `OutboundJaxrsResponse.from()`. It
just copies all headers as-is, and the headers have not been filtered
earlier in the process either.

This PR implements a helper in HeaderUtils (+ unit test) that copies
only non-hop-by-hop headers and takes it into use in
`OutboundJaxrsResponse.from()` (which imo is the correct place to do
it).

Signed-off-by: Jonas Berlin <xkr47@outerspace.dyndns.org>